Move data to a separate opt-test-data crate [1/2]
Since this crate is dedicated to data, we embed it using include_dir to
remove all complications of locating it in the filesystem.  A function
is dedicated for copying the data into a target dir.

opr crate then gets the data though build.rs.

This is an intermediate step that adds a build.rs only needed for
test, which impacts normal builds.  Will be addressed in next commit
for clarity.