Add "local-files" feature for no-network testing operation
Army URL is not fixed any more, introduce get_army_url().

Intended for local testing of web apps using data from opr-test-data,
intended to be accessible from a local webserver in /data/, as installed by
`opr_test_data::import_data()`.  Since the actual location depends on the
location of the data files to serve it is the responsibility of the app to
get those files installed in the proper location.

For `trunk serve` for example, data files have to end up being installed
as dist/data/.