Army: deserialize upgrade costs
Since the unit cost we have is the base cost, we have to collect the
upgrade costs to get useful values.

Unfortunately, `selectedUpgrades[].option.cost` is legacy/wrong and we
have to look the actual cost in `costs`, so deserialize it as a
HashMap.  We already extract the Unit::id necessary to make the lookup.